rename route from scicat
