+++ src/server/routes.ts
@@ -23,12 +23,12 @@
  */
 
 import { Application } from 'express';
-import * as scicat from "../providers/controllers/scicat";
+import * as scicat from "../providers/controllers/plexus";
 import * as openaire from "../providers/controllers/openaire";
 import logger from "./logger";
 
 export default function routes(app: Application): void {
   logger.debug('Setting express routes for OAI providers.');
-  app.get('/scicat/oai', scicat.oai('/scicat/oai'));
+  app.get('/plexus/oai', scicat.oai('/plexus/oai'));
   app.get('/openaire/oai', openaire.oai('/openaire/oai'));
 };
